[RHDM-709][RHPAM-1453] upgrade mvel version (#2196)
