+++ drools-compiler/src/main/java/org/drools/compiler/rule/builder/dialect/asm/InvokerGenerator.java
@@ -15,6 +15,10 @@
 
 package org.drools.compiler.rule.builder.dialect.asm;
 
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Set;
+
 import org.drools.compiler.rule.builder.RuleBuildContext;
 import org.drools.core.rule.Declaration;
 import org.drools.core.rule.builder.dialect.asm.ClassGenerator;
@@ -22,10 +26,6 @@ import org.drools.core.rule.builder.dialect.asm.GeneratorHelper;
 import org.drools.core.rule.builder.dialect.asm.InvokerDataProvider;
 import org.kie.soup.project.datamodel.commons.types.TypeResolver;
 import org.mvel2.asm.MethodVisitor;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
 
 import static org.mvel2.asm.Opcodes.ACC_FINAL;
 import static org.mvel2.asm.Opcodes.ACC_PRIVATE;
@@ -58,7 +58,7 @@ public class InvokerGenerator {
                 push(data.hashCode());
                 mv.visitInsn(IRETURN);
             }
-        }).addMethod(ACC_PUBLIC, "getMethodBytecode", generator.methodDescr(List.class), new GeneratorHelper.GetMethodBytecodeMethod(data)
+        }).addMethod(ACC_PUBLIC, "getMethodBytecode", generator.methodDescr(String.class), new GeneratorHelper.GetMethodBytecodeMethod(data)
         ).addMethod(ACC_PUBLIC, "equals", generator.methodDescr(Boolean.TYPE, Object.class), new GeneratorHelper.EqualsMethod()
         ).addMethod(ACC_PUBLIC, "getPackageName", generator.methodDescr(String.class), new ClassGenerator.MethodBody() {
             public void body(MethodVisitor mv) {
