+++ drools-test-coverage/test-suite/src/test/java/org/drools/testcoverage/regression/mvel/MvelLinkageErrorTest.java
@@ -16,16 +16,14 @@
 
 package org.drools.testcoverage.regression.mvel;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import org.assertj.core.api.Assertions;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 import org.kie.api.KieServices;
 import org.kie.api.builder.KieFileSystem;
 import org.kie.api.runtime.KieSession;
-
-import java.util.ArrayList;
-import java.util.List;
 
 /**
  * Tests handling a collision of identifer and dynamic package import (BZ 1321281).
@@ -46,37 +44,22 @@ public class MvelLinkageErrorTest {
             + "    output.add(NotLoadableClass); \n"
             + "end\n";
 
-    private KieSession kieSession;
-
-    @Before
-    public void init() {
-        KieServices ks = KieServices.Factory.get();
-        KieFileSystem kfs = ks.newKieFileSystem().write("src/main/resources/r1.drl", DRL);
-        ks.newKieBuilder( kfs ).buildAll();
-
-        kieSession = ks.newKieContainer(ks.getRepository().getDefaultReleaseId()).newKieSession();
-    }
-
-    @After
-    public void cleanup() {
-        if (kieSession != null) {
-            kieSession.dispose();
-        }
-    }
-
     @Test
-    public void testMvelLinkageError() throws Exception {
-        final List<Integer> result = new ArrayList<Integer>();
-        kieSession.setGlobal("output", result);
+    public void testMvelLinkageError() {
+        final KieServices ks = KieServices.Factory.get();
+        final KieFileSystem kfs = ks.newKieFileSystem().write("src/main/resources/r1.drl", DRL);
+        Assertions.assertThatCode(() -> ks.newKieBuilder(kfs).buildAll()).doesNotThrowAnyException();
 
-        kieSession.insert(TEST_FACT);
+        final KieSession kieSession = ks.newKieContainer(ks.getRepository().getDefaultReleaseId()).newKieSession();
         try {
-            int fired = kieSession.fireAllRules();
+            final List result = new ArrayList<>();
+            kieSession.setGlobal("output", result);
+
+            kieSession.insert(TEST_FACT);
+            final int fired = kieSession.fireAllRules();
             Assertions.assertThat(fired).as("Unexpected number of rules fired.").isEqualTo(1);
-            Assertions.assertThat(result).as("Rule produced unexpected result value.").containsExactly(TEST_FACT.length());
         } finally {
             kieSession.dispose();
         }
     }
-
 }
